WIP: Polish benefits componen and page styles
